Feat/add title routes

+++ src/domain/titleDomain.ts
@@ -7,13 +7,14 @@ export const TitleDomain = {
     author,
     finished,
     publisher: { name: publisherName = "", id: publisherId },
-    ...titleDTOProps
+    releasedVolumes,
   }: TitleDTO) => {
     const createdTitle = await TitleRepository.CreateTitle({
       name,
       author,
       finished,
       publisher: { name: publisherName, id: publisherId },
+      releasedVolumes,
     });
     return createdTitle;
   },
@@ -32,17 +33,23 @@ export const TitleDomain = {
       publisher: { name: publisherName, id: publisherId },
       id,
     };
-    const updatedUser = await TitleRepository.UpdateTitle(uptadeUser);
-    return updatedUser;
+    const updatedTitle = await TitleRepository.UpdateTitle(uptadeUser);
+    return updatedTitle;
   },
 
   DeleteTitle: async (titleId: string) => {
-    const deletedUser = await TitleRepository.DeleteTitle(titleId);
-    return deletedUser;
+    if (titleId) {
+      const deletedTitle = await TitleRepository.DeleteTitle(titleId);
+      return deletedTitle;
+    }
+    throw new Error("id inválido");
   },
 
   GetTitle: async (titleId: string) => {
-    const title = await TitleRepository.GetTitle(titleId);
-    return title;
+    if (titleId) {
+      const title = await TitleRepository.GetTitle(titleId);
+      return title;
+    }
+    throw new Error("id inválido");
   },
 };
